+++ api/src/Fixture/CaseStatus.php
@@ -25,16 +25,6 @@ trait CaseStatus
      */
     public function load(ObjectManager $manager)
     {
-        $connection = $manager->getConnection();
-        $platform = $connection->getDatabasePlatform()->getName();
-
-        switch ($platform) {
-            case 'postgresql':
-                $connection->exec('ALTER SEQUENCE app_case_status_id_seq RESTART WITH 1');
-                $connection->exec('ALTER SEQUENCE app_case_status_trans_id_seq RESTART WITH 1');
-                break;
-        }
-
         $objects = $this->parse($this->path);
 
         foreach ($objects as $object) {