+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.15.0
 
 - Feature [Composer] Upgrade apip dependency to 2.3.4
+- Feature [Fixture] Separate sequence reset fixture from entity fixture
 - Feature [App] Case Status entity path changed from `case-statuses` to `case_statuses`
 
 ## 0.14.1 (2018-11-06)